First round of refactoring
